Crate location_utils

Utilities for finding and comparing longitude- and latitude-based locations using S2 Geometry.

See https://s2geometry.io/ for more information.

Structs

Angle

Angle represents a 1D angle. The internal representation is a double precision value in radians, so conversion to and from radians is exact. Conversions between E5, E6, E7, and Degrees are not always exact. For example, Deg(3.1) is different from E6(3100000) or E7(310000000).

CellID

CellID uniquely identifies a cell in the S2 cell decomposition. The most significant 3 bits encode the face number (0-5). The remaining 61 bits encode the position of the center of this cell along the Hilbert curve on that face. The zero value and the value (1<<64)-1 are invalid cell IDs. The first compares less than any valid cell ID, the second as greater than any valid cell ID.

LatLng

Constants

DEFAULT_CUTOFF_RADIUS_RADIANS

The default cutoff radius in radians.

LOCATION_SIZE

S2_DEFAULT_LEVEL

The default level to use for S2 Geometry cells. Higher numbers means smaller cells. At level 7 the cells have roughly similar size to a circle with a radius of 40km and there are apporimately 98K of them.

Traits

Context

Provides the context method for Result.

Functions

cell_id_from_bytes

Converts a byte representation to the CellID.

cell_id_to_bytes

Converts a CellID to a byte representation.

find_cell

Finds the cell at the given level that covers the location.

find_covering_cells

Finds the set of S2 cells at the required level that covers the area around the location. The area is a spherical cap with the location as the midpoint with the specified radius.

location_from_bytes

Converts a byte representation of a location into a LatLng.

location_from_degrees

Converts latitude and longitude values in degrees to a LatLng location.

location_to_bytes

Converts a LatLng location to a byte representation.

Type Definitions

Result

Result<T, Error>