Function location_utils::find_covering_cells

pub fn find_covering_cells(
    location: &LatLng, 
    radius: &Angle, 
    level: u8
) -> Result<Vec<CellID>>

Finds the set of S2 cells at the required level that covers the area around the location. The area is a spherical cap with the location as the midpoint with the specified radius.

The level determines the size of the cells. We use a single fixed level for the covering (min level = max level). If the level is higher (and therefore the cells smaller) the coverage is usually tighter with less area that falls inside the cells but outside of the target zone. If the level is lower it would usually require fewer cells to cover the zone, but there is usually larger areas that fall outside of the zone.

See https://s2geometry.io/devguide/examples/coverings for more information on coverings.